back-end/requirements.txt
boto3
python-dotenv
flask
flask-cors
gunicorn